# tmf639_openapi_spec.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Resource Inventory Management",
        "description": "## TMF API Reference: TMF639 - Resource Inventory \n\n### Release : 19.5 - December 2019\n\nResource Inventory  API goal is to provide the ability to manage Resources.\n\n### Operations\nResource Inventory API performs the following operations on the resources :\n- Retrieve an entity or a collection of entities depending on filter criteria\n- Partial update of an entity (including updating rules)\n- Create an entity (including default values and creation rules)\n- Delete an entity (for administration purposes)\n- Manage notification of events",
        "version": "4.0.0"
    },
    "host": "serverRoot",
    "basePath": "/tmf-api/resourceInventoryManagement/v4/",
    "schemes": [
        "https"
    ],
    "consumes": [
        "application/json;charset=utf-8"
    ],
    "produces": [
        "application/json;charset=utf-8"
    ],
...
    "paths": {
        "/resource": {
            "get": {
                "operationId": "listResource",
                "summary": "List or find Resource objects",
                "description": "This operation list or find Resource entities",
                "tags": [
                    "resource"
                ],
                "parameters": [
                    {
                        "name": "fields",
                        "description": "Comma-separated properties to be provided in response",
                        "required": false,
                        "in": "query",
                        "type": "string"
                    },
                    {
                        "name": "offset",
                        "description": "Requested index for start of resources to be provided in response",
                        "required": false,
                        "in": "query",
                        "type": "integer"
                    },
                    {
                        "name": "limit",
                        "description": "Requested number of resources to be provided in response",
                        "required": false,
                        "in": "query",
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "headers": {
                            "X-Result-Count": {
                                "description": "Actual number of items returned in the response body",
                                "type": "integer"
                            },
                            "X-Total-Count": {
                                "description": "Total number of items matching criteria",
                                "type": "integer"
                            }
                        },
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Resource"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "405": {
                        "description": "Method Not allowed",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "post": {
                "operationId": "createResource",
                "summary": "Creates a Resource",
                "description": "This operation creates a Resource entity.",
                "tags": [
                    "resource"
                ],
                "parameters": [
                    {
                        "name": "resource",
                        "description": "The Resource to be created",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Resource_Create"
                        },
                        "in": "body"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/Resource"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "405": {
                        "description": "Method Not allowed",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/resource/{id}": {
            "get": {
                "operationId": "retrieveResource",
                "summary": "Retrieves a Resource by ID",
                "description": "This operation retrieves a Resource entity. Attribute selection is enabled for all first level attributes.",
                "tags": [
                    "resource"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "description": "Identifier of the Resource",
                        "required": true,
                        "type": "string",
                        "in": "path"
                    },
                    {
                        "name": "fields",
                        "description": "Comma-separated properties to provide in response",
                        "required": false,
                        "type": "string",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success",
                        "schema": {
                            "$ref": "#/definitions/Resource"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "405": {
                        "description": "Method Not allowed",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "patch": {
                "operationId": "patchResource",
                "summary": "Updates partially a Resource",
                "description": "This operation updates partially a Resource entity.",
                "tags": [
                    "resource"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "description": "Identifier of the Resource",
                        "required": true,
                        "type": "string",
                        "in": "path"
                    },
                    {
                        "name": "resource",
                        "description": "The Resource to be updated",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/Resource_Update"
                        },
                        "in": "body"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Updated",
                        "schema": {
                            "$ref": "#/definitions/Resource"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "405": {
                        "description": "Method Not allowed",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            },
            "delete": {
                "operationId": "deleteResource",
                "summary": "Deletes a Resource",
                "description": "This operation deletes a Resource entity.",
                "tags": [
                    "resource"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "description": "Identifier of the Resource",
                        "required": true,
                        "type": "string",
                        "in": "path"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Deleted"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "405": {
                        "description": "Method Not allowed",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Feature": {
            "type": "object",
            "description": "Configuration feature.",
            "required": [
                "featureCharacteristic",
                "name"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Unique identifier of the feature."
                },
                "isBundle": {
                    "type": "boolean",
                    "description": "True if this is a feature group. Default is false."
                },
                "isEnabled": {
                    "type": "boolean",
                    "description": "True if this feature is enabled. Default is true."
                },
                "name": {
                    "type": "string",
                    "description": "This is the name for the feature."
                },
                "constraint": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ConstraintRef"
                    },
                    "description": "This is a list of feature constraints."
                },
                "featureCharacteristic": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Characteristic"
                    },
                    "minItems": 1,
                    "description": "This is a list of Characteristics for a particular feature."
                },
                "featureRelationship": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/FeatureRelationship"
                    }
                }
            }
        },
...
        "Resource": {
            "type": "object",
            "description": "Resource is an abstract entity that describes the common set of attributes shared by all concrete resources (e.g. TPE, EQUIPMENT) in the inventory.",
            "required": [
                "href",
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string",
                    "description": "Identifier of an instance of the resource. Required to be unique within the resource type.  Used in URIs as the identifier for specific instances of a type."
                },
                "href": {
                    "type": "string",
                    "description": "The URI for the object itself."
                },
                "category": {
                    "type": "string",
                    "description": "Category of the concrete resource. e.g Gold, Silver for MSISDN concrete resource"
                },
                "description": {
                    "type": "string",
                    "description": "free-text description of the resource"
                },
                "endOperatingDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "A date time( DateTime). The date till the resource is operating"
                },
                "name": {
                    "type": "string",
                    "description": "A string used to give a name to the resource"
                },
                "resourceVersion": {
                    "type": "string",
                    "description": "A field that identifies the specific version of an instance of a resource."
                },
                "startOperatingDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "A date time( DateTime). The date from which the resource is operating"
                },
                "activationFeature": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Feature"
                    },
                    "description": "Configuration features"
                },
                "administrativeState": {
                    "$ref": "#/definitions/ResourceAdministrativeStateType",
                    "description": "Tracks the lifecycle status of the resource, such as planning, installing, opereating, retiring and so on."
                },
                "attachment": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/AttachmentRefOrValue"
                    }
                },
                "note": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Note"
                    }
                },
                "operationalState": {
                    "$ref": "#/definitions/ResourceOperationalStateType",
                    "description": "Tracks the lifecycle status of the resource, such as planning, installing, opereating, retiring and so on."
                },
                "place": {
                    "$ref": "#/definitions/RelatedPlaceRefOrValue"
                },
                "relatedParty": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RelatedParty"
                    }
                },
                "resourceCharacteristic": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Characteristic"
                    }
                },
                "resourceRelationship": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResourceRelationship"
                    }
                },
                "resourceSpecification": {
                    "$ref": "#/definitions/ResourceSpecificationRef"
                },
                "resourceStatus": {
                    "$ref": "#/definitions/ResourceStatusType",
                    "description": "Tracks the lifecycle status of the resource, such as planning, installing, opereating, retiring and so on."
                },
                "usageState": {
                    "$ref": "#/definitions/ResourceUsageStateType",
                    "description": "Tracks the lifecycle status of the resource, such as planning, installing, opereating, retiring and so on."
                },
                "@baseType": {
                    "type": "string",
                    "description": "When sub-classing, this defines the super-class"
                },
                "@schemaLocation": {
                    "type": "string",
                    "format": "uri",
                    "description": "A URI to a JSON-Schema file that defines additional attributes and relationships"
                },
                "@type": {
                    "type": "string",
                    "description": "When sub-classing, this defines the sub-class Extensible name"
                }
            }
        },
        "Resource_Create": {
            "type": "object",
            "description": "Resource is an abstract entity that describes the common set of attributes shared by all concrete resources (e.g. TPE, EQUIPMENT) in the inventory.\nSkipped properties: id,href",
            "required": [
                "href",
                "id",
                "name"
            ],
            "properties": {
                "category": {
                    "type": "string",
                    "description": "Category of the concrete resource. e.g Gold, Silver for MSISDN concrete resource"
                },
                "description": {
                    "type": "string",
                    "description": "free-text description of the resource"
                },
                "endOperatingDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "A date time( DateTime). The date till the resource is operating"
                },
                "name": {
                    "type": "string",
                    "description": "A string used to give a name to the resource"
                },
                "resourceVersion": {
                    "type": "string",
                    "description": "A field that identifies the specific version of an instance of a resource."
                },
                "startOperatingDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "A date time( DateTime). The date from which the resource is operating"
                },
                "activationFeature": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Feature"
                    },
                    "description": "Configuration features"
                },
                "administrativeState": {
                    "$ref": "#/definitions/ResourceAdministrativeStateType",
                    "description": "Tracks the lifecycle status of the resource, such as planning, installing, opereating, retiring and so on."
                },
                "attachment": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/AttachmentRefOrValue"
                    }
                },
                "note": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Note"
                    }
                },
                "operationalState": {
                    "$ref": "#/definitions/ResourceOperationalStateType",
                    "description": "Tracks the lifecycle status of the resource, such as planning, installing, opereating, retiring and so on."
                },
                "place": {
                    "$ref": "#/definitions/RelatedPlaceRefOrValue"
                },
                "relatedParty": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RelatedParty"
                    }
                },
                "resourceCharacteristic": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Characteristic"
                    }
                },
                "resourceRelationship": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResourceRelationship"
                    }
                },
                "resourceSpecification": {
                    "$ref": "#/definitions/ResourceSpecificationRef"
                },
                "resourceStatus": {
                    "$ref": "#/definitions/ResourceStatusType",
                    "description": "Tracks the lifecycle status of the resource, such as planning, installing, opereating, retiring and so on."
                },
                "usageState": {
                    "$ref": "#/definitions/ResourceUsageStateType",
                    "description": "Tracks the lifecycle status of the resource, such as planning, installing, opereating, retiring and so on."
                },
                "@baseType": {
                    "type": "string",
                    "description": "When sub-classing, this defines the super-class"
                },
                "@schemaLocation": {
                    "type": "string",
                    "format": "uri",
                    "description": "A URI to a JSON-Schema file that defines additional attributes and relationships"
                },
                "@type": {
                    "type": "string",
                    "description": "When sub-classing, this defines the sub-class Extensible name"
                }
            }
        },
        "Resource_Update": {
            "type": "object",
            "description": "Resource is an abstract entity that describes the common set of attributes shared by all concrete resources (e.g. TPE, EQUIPMENT) in the inventory.\nSkipped properties: id,href",
            "required": [
                "href",
                "id"
            ],
            "properties": {
                "category": {
                    "type": "string",
                    "description": "Category of the concrete resource. e.g Gold, Silver for MSISDN concrete resource"
                },
                "description": {
                    "type": "string",
                    "description": "free-text description of the resource"
                },
                "endOperatingDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "A date time( DateTime). The date till the resource is operating"
                },
                "name": {
                    "type": "string",
                    "description": "A string used to give a name to the resource"
                },
                "resourceVersion": {
                    "type": "string",
                    "description": "A field that identifies the specific version of an instance of a resource."
                },
                "startOperatingDate": {
                    "type": "string",
                    "format": "date-time",
                    "description": "A date time( DateTime). The date from which the resource is operating"
                },
                "activationFeature": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Feature"
                    },
                    "description": "Configuration features"
                },
                "administrativeState": {
                    "$ref": "#/definitions/ResourceAdministrativeStateType",
                    "description": "Tracks the lifecycle status of the resource, such as planning, installing, opereating, retiring and so on."
                },
                "attachment": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/AttachmentRefOrValue"
                    }
                },
                "note": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Note"
                    }
                },
                "operationalState": {
                    "$ref": "#/definitions/ResourceOperationalStateType",
                    "description": "Tracks the lifecycle status of the resource, such as planning, installing, opereating, retiring and so on."
                },
                "place": {
                    "$ref": "#/definitions/RelatedPlaceRefOrValue"
                },
                "relatedParty": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/RelatedParty"
                    }
                },
                "resourceCharacteristic": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Characteristic"
                    }
                },
                "resourceRelationship": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/ResourceRelationship"
                    }
                },
                "resourceSpecification": {
                    "$ref": "#/definitions/ResourceSpecificationRef"
                },
                "resourceStatus": {
                    "$ref": "#/definitions/ResourceStatusType",
                    "description": "Tracks the lifecycle status of the resource, such as planning, installing, opereating, retiring and so on."
                },
                "usageState": {
                    "$ref": "#/definitions/ResourceUsageStateType",
                    "description": "Tracks the lifecycle status of the resource, such as planning, installing, opereating, retiring and so on."
                },
                "@baseType": {
                    "type": "string",
                    "description": "When sub-classing, this defines the super-class"
                },
                "@schemaLocation": {
                    "type": "string",
                    "format": "uri",
                    "description": "A URI to a JSON-Schema file that defines additional attributes and relationships"
                },
                "@type": {
                    "type": "string",
                    "description": "When sub-classing, this defines the sub-class Extensible name"
                }
            }
        },
...
        "Error": {
            "description": "Used when an API throws an Error, typically with a HTTP error response-code (3xx, 4xx, 5xx)",
            "type": "object",
            "required": [
                "code",
                "reason"
            ],
            "properties": {
                "code": {
                    "type": "string",
                    "description": "Application relevant detail, defined in the API or a common list."
                },
                "reason": {
                    "type": "string",
                    "description": "Explanation of the reason for the error which can be shown to a client user."
                },
                "message": {
                    "type": "string",
                    "description": "More details and corrective actions related to the error which can be shown to a client user."
                },
                "status": {
                    "type": "string",
                    "description": "HTTP Error code extension"
                },
                "referenceError": {
                    "type": "string",
                    "format": "uri",
                    "description": "URI of documentation describing the error."
                },
                "@baseType": {
                    "type": "string",
                    "description": "When sub-classing, this defines the super-class."
                },
                "@schemaLocation": {
                    "type": "string",
                    "format": "uri",
                    "description": "A URI to a JSON-Schema file that defines additional attributes and relationships"
                },
                "@type": {
                    "type": "string",
                    "description": "When sub-classing, this defines the sub-class entity name."
                }
            }
        }
    }
}